Fixes flaky integration tests.

+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/ConfigMapIT.java
@@ -16,6 +16,8 @@
 
 package io.fabric8.kubernetes;
 
+import io.fabric8.commons.DeleteEntity;
+import io.fabric8.commons.ReadyEntity;
 import io.fabric8.kubernetes.api.model.*;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import org.arquillian.cube.kubernetes.api.Session;
@@ -27,8 +29,11 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import java.util.concurrent.TimeUnit;
+
 import static junit.framework.TestCase.assertTrue;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.awaitility.Awaitility.await;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
@@ -89,10 +94,17 @@ public class ConfigMapIT {
 
   @Test
   public void update() {
+    ReadyEntity<ConfigMap> configMap1Ready = new ReadyEntity<>(ConfigMap.class, client, "configmap1", currentNamespace);
+    ReadyEntity<ConfigMap> configMap2Ready = new ReadyEntity<>(ConfigMap.class, client, "configmap2", currentNamespace);
+
     configMap1 = client.configMaps().inNamespace(currentNamespace).withName("configmap1").edit()
       .addToData("4", "four").done();
+
     configMap2 = client.configMaps().inNamespace(currentNamespace).withName("configmap2").edit()
       .addToData("MSSQL", "Microsoft Database").done();
+
+    await().atMost(30, TimeUnit.SECONDS).until(configMap1Ready);
+    await().atMost(30, TimeUnit.SECONDS).until(configMap2Ready);
     assertNotNull(configMap1);
     assertNotNull(configMap2);
     assertEquals("four", configMap1.getData().get("4"));
@@ -101,14 +113,21 @@ public class ConfigMapIT {
 
   @Test
   public void delete() {
+    ReadyEntity<ConfigMap> configMap1Ready = new ReadyEntity<>(ConfigMap.class, client, "configmap1", currentNamespace);
+    await().atMost(30, TimeUnit.SECONDS).until(configMap1Ready);
     assertTrue(client.configMaps().inNamespace(currentNamespace).withName("configmap1").delete());
+
+    ReadyEntity<ConfigMap> configMap2Ready = new ReadyEntity<>(ConfigMap.class, client, "configmap2", currentNamespace);
+    await().atMost(30, TimeUnit.SECONDS).until(configMap2Ready);
     assertTrue(client.configMaps().inNamespace(currentNamespace).withName("configmap2").delete());
   }
 
   @After
   public void cleanup() throws InterruptedException {
-    client.configMaps().inNamespace(currentNamespace).delete();
-    // Wait for resources to get destroyed
-    Thread.sleep(30000);
+    if (client.configMaps().inNamespace(currentNamespace).list().getItems().size()!= 0) {
+      client.configMaps().inNamespace(currentNamespace).delete();
+    }
+    DeleteEntity<ConfigMap> configMapDelete = new DeleteEntity<>(ConfigMap.class, client, "configmap1", currentNamespace);
+    await().atMost(30, TimeUnit.SECONDS).until(configMapDelete);
   }
 }
